backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn",
    "pyyaml",
]
